+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - allow one skill type per user
 - date_last_txp_add db column under skill
 - restrict adding txp more than time diff of last add date
+- custom `migrate.go` script to load sql migrations into db
 
 ### Changed
 ### Fixed